[2.4sp] How many?
model solution - for the unique elements of the array, perform a count
of each and add element/count to hash as key and value respectively.
Having done this, for each pair in the "count" hash, print the pair.

group_by solution - this tends to be what you find when you search for
"count things and put them in a hash". group_by is interesting, here is
the description from the docs:

group_by { |obj| block } → a_hash
Returns a hash, so this is a array to hash mapping (unless you don't
include a block, and then an enumerator is returned.) Good sign for this
problem, a hash return value should be easy to get the output desired.
It also helps to know that a hash is a common idiom for this kind of
problem, keys are the things counted, values are the count.

"Groups the collection by result of the block." Somewhat vague, but
result of block is important. As it relates to the problem, at a high
level we want to group the elements in the array and count them.

"Returns a hash where the keys are the evaluated result from the block"
So the keys of the returned hash are the evalution of each element in
the array when put through the expression contained in the block.

"and the values are arrays of elements in the collection that correspond
to the key." Note the values of the returned hash are arrays.
"Correspond to the key" is intuitive but vague. Recall the keys in the
returned hash are the result of evaluating the block, so the
correspondence is: collect those elements from the argument array that
mapped to a particular key, collect them into an array, and this is the
value for that particular key. In math function terms, the key is a
point in the image of the block expression/evaluation, and the values
are this the inverse image of its key.

Image/Inverse-Image => https://en.wikipedia.org/wiki/Image_(mathematics)

Sooooo..... we passed the "identity" block meaning each array element is
evaluated to itself. Thus the hash results in each unique element being
a key in the group_by hash, and its value is an array with n copies of
the element -- the inverse image of the identity map is each of the
instances of the element in the argument array. N.B.  Kernel#itself was
added to Ruby 2.2, so group_by(&:itself) will pass the identity to
group_by.

SO discussion about itself:
https://stackoverflow.com/questions/16932711/is-there-a-short-way-to-write-x-x

N.N.B. After further invenstigation, itself is a method on Object (in
2.7 at least), not kernel.

Hash#transform_values is like map on just the values in a hash, take
each value and replace it with the value's evaluation from the block.
Here, the values are of type array, so we call size on them to get the
count.

tally solution - tally is a nice convenience method added in 2.7, it has
the exact functionality we want, make a hash with the counts of the
elements of the argument array.

case-insensitive solution - for the model solution, apply either upcase
or downcase prior to applying uniq to the argument array. For the
group_by, the block expression should be down/up casing, and for the
tally solution you can map the array with up/down case.